% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{annotateVariant}
\alias{annotateVariant}
\title{Variant annotation}
\usage{
annotateVariant(variant.names)
}
\arguments{
\item{variant.names}{input variant IDs, can be a vector.}
}
\value{
a data frame with annotation for all input variant IDs.
}
\description{
This function takes variant names as input and
returns annotation for annotation table for all variant IDs in a data frame.
}
\examples{
annotateVariant(variants_id)

}
\keyword{variant}
