% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\name{elbowPlot}
\alias{elbowPlot}
\title{Generate elbow plot}
\usage{
elbowPlot(input.mtx)
}
\arguments{
\item{input.mtx}{Input matrix}
}
\value{
an elbow plot
}
\description{
This function generate an elbow plot. The plot is useful for determining how many PCs will be used for dimension reduction
The input matrix can be the DNA matrix in an optima object.
}
\examples{
elbowPlot(example.matrix)
}
