% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{filterVariant}
\alias{filterVariant}
\title{DNA Variant filter function}
\usage{
filterVariant(
  optima.obj,
  min.dp = 10,
  min.gq = 30,
  vaf.ref = 5,
  vaf.hom = 95,
  vaf.het = 35,
  min.cell.pt = 50,
  min.mut.cell.pt = 1
)
}
\arguments{
\item{optima.obj}{An optima object with raw data unfiltered.}

\item{min.dp}{Minimum depth, defaults to 10.}

\item{min.gq}{Minimum genotype quality, defaults to 30.}

\item{vaf.ref}{If reference call vaf (GT=0) is larger than vaf.ref, then value in genotype call matrix is converted to GT=3.}

\item{vaf.hom}{If homozygous call vaf (GT=2) is smaller than vaf.hom, then value in genotype call matrix is converted to GT=3.}

\item{vaf.het}{If heterozygous call vaf (GT=1) is smaller than vaf.ref, then value in genotype call matrix is converted to GT=3.}

\item{min.cell.pt}{Minimum threshold for cell percentage that has valid variant call (GT = 0, 1 or 2) after applying the filter.}

\item{min.mut.cell.pt}{Minimum threshold for cell percentage that has mutated genotype (GT = 1 or 2) after applying the filter.}
}
\value{
An optima object, The DNA data in the object is filtered, the variant.filter label is "filtered".
Meanwhile, the protein matrix and CNV matrix is also updated so that only cells withstand DNA variant filter are kept.
}
\description{
This function uses multiple matrices imported from the h5 file to conduct quality filtering.
This includes the sequencing depth matrix, genotype matrix, variant allele frequency matrix, genotype quality matrix
The function returns an optima object that has been filtered with variant/cells.
In addition, the returned optima object's variant.filter label is changed to "filtered".
This function is usually applied before protein and CNV analysis.
}
\examples{
filterVariant(optima.obj)
}
\keyword{DNA}
\keyword{filter}
