% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{getClones}
\alias{getClones}
\title{Clustering variant function}
\usage{
getClones(optima.obj, eps = 1, minPts = 100, num.PC = 5, plot = FALSE)
}
\arguments{
\item{optima.obj}{optima object.}

\item{eps}{size/radius of the epsilon neighborhood.
This argument will passed to dbscan function.}

\item{minPts}{number of minimum points required in the eps neighborhood
for core points, including the point itself.
This argument will passed to dbscan function.}

\item{plot}{if True, a UMAP plot will be generated based on the dimension reduction
result from variant matrix.
Default is FALSE.}
}
\value{
A list, the first element in the list is the optima
object with labels assigned based on dbscan. the second element
in the list is the dimension reduction result based on VAF matrix.
}
\description{
This function identifies cell clones based on DNA variant data.
}
\examples{
getClones(my.obj)
}
