% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{getCNVmtx}
\alias{getCNVmtx}
\title{The getter function for CNV matrix}
\usage{
getCNVmtx(optima.obj)
}
\arguments{
\item{optima.obj}{optima object.}
}
\value{
A matrix that contains CNV data in the optima object.
The row names are cell id, the column names are CNV IDs.
}
\description{
This function returns the CNV matrix within the optima object.
}
\examples{
getCNVmtx(my.obj)
}
