% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{getDNAmtx}
\alias{getDNAmtx}
\title{The getter function for VAF matrix}
\usage{
getDNAmtx(optima.obj)
}
\arguments{
\item{optima.obj}{optima object.}
}
\value{
A matrix that contains VAF data in the optima object.
The row names are cell id, the column names are variant ID.
}
\description{
This function returns the VAF matrix within the optima object.
}
\examples{
getDNAmtx(my.obj)
}
