% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{getInfo}
\alias{getInfo}
\title{Single variant ID annotation function}
\usage{
getInfo(variant)
}
\arguments{
\item{variant}{variant name in a string}
}
\value{
Annotation information for one specific variant ID.
}
\description{
This function returns the annotation from one variant ID. This function is not visable to users.
}
\examples{
getInfo(variant_id)
}
\keyword{variant}
