% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{getPloidyMtx}
\alias{getPloidyMtx}
\title{The getter function for Ploidy matrix}
\usage{
getPloidyMtx(optima.obj)
}
\arguments{
\item{optima.obj}{optima object.}
}
\value{
A matrix that contains Ploidy data in the optima object.
The row names are cell id, the column names are CNV IDs.
}
\description{
This function returns the Ploidy matrix within the optima object.
}
\examples{
getPloidyMtx(my.obj)
}
