% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{getProteinMtx}
\alias{getProteinMtx}
\title{The getter function for Protein matrix}
\usage{
getProteinMtx(optima.obj)
}
\arguments{
\item{optima.obj}{optima object.}
}
\value{
A matrix that contains Protein data in the optima object.
The row names are cell IDs, the column names are protein IDs.
}
\description{
This function returns the Protein matrix within the optima object.
}
\examples{
getProteinMtx(my.obj)
}
