% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{normalizeCNV}
\alias{normalizeCNV}
\title{CNV normalization function}
\usage{
normalizeCNV(optima.obj)
}
\arguments{
\item{optima.obj}{optima object}
}
\value{
optima object with normalized CNV and amp.normalize.method updated to "normalized"
}
\description{
The function normalizes the CNV matrix to correct for column-wise and row-wise variation and
updates the optima object amp.normalize.method from "unnormalized" to "normalized".
}
\examples{
normalizeCNV(optima.obj)
}
\keyword{optima.obj}
