% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{normalizeProtein}
\alias{normalizeProtein}
\title{Protein matrix normalization}
\usage{
normalizeProtein(optima.object)
}
\arguments{
\item{optima.obj}{an optima object with raw data unfiltered}
}
\value{
An optima object with protein matrix being normalized and
protein.normalize.method label updated to "normalized"
}
\description{
The function normalizes protein matrix within an optima object using
CLR method.
}
\examples{
normalizeProtein(optima.object)
}
\keyword{optima.obj}
