% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\docType{class}
\name{optima-class}
\alias{optima-class}
\alias{optima}
\title{optima object}
\value{
Object containing DNA, protein and CNV single cell sequencing data.
}
\description{
An optima object contains DNA, protein and CNV for Tapestri platform
single cell sequencing data.
}
\section{Slots}{

\describe{
\item{\code{meta.data}}{User-defined metadata can be kept with the object.}

\item{\code{cell.ids}}{A vector of cell IDs/barcodes from Tapestri. This
vector should contain unique IDs.}

\item{\code{cell.labels}}{A vector that is used to store the cell type information
for each cell.}

\item{\code{variants}}{A vector of variant IDs.}

\item{\code{variant.filter}}{A string that keeps track of if optima object is being QC filtered
on its variant matrix.}

\item{\code{vaf.mtx}}{Variant matrix.}

\item{\code{gt.mtx}}{Genotype matrix.}

\item{\code{dp.mtx}}{Sequencing depth matrix.}

\item{\code{gq.mtx}}{Genotype quality.}

\item{\code{amps}}{A vector of CNV locus.}

\item{\code{amp.normalized.method}}{A string that keeps track of if optima object is being normalized
on its CNV matrix.}

\item{\code{amp.mtx}}{CNV matrix.}

\item{\code{ploidy.mtx}}{Ploidy matrix.}

\item{\code{proteins}}{A vector of surface protein ID.}

\item{\code{protein.normalize.method}}{A string that keeps track of if optima object is being normalized
on its protein matrix.}

\item{\code{protein.mtx}}{Protein matrix}
}}

\examples{
setClass()
}
