% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{plotPloidy}
\alias{plotPloidy}
\title{Ploidy scatter plot function}
\usage{
plotPloidy(optima.obj, cell.type)
}
\arguments{
\item{optima.obj}{optima object.}

\item{cell.type}{String that indicates which cell type,}
}
\value{
optima object with normalized CNV.
}
\description{
For a specified cell type, this function creates a scatter plot indicating
ploidy for different CNV loci.
}
\examples{
plotPloidy()
}
\keyword{optima.obj}
