% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{plotProteinFeature}
\alias{plotProteinFeature}
\title{Plot specific protein levels}
\usage{
plotProteinFeature(optima.obj, protein.name, reduceDim.obj)
}
\arguments{
\item{optima.obj}{optima object.}

\item{protein.name}{the specific protein user interested}

\item{reduceDim.obj}{dimension reduction result returned by reduceDim() function.}
}
\value{
A scatter plot based on dimension reduction result. Each cell is colored
based on the protein level. The more expression.
}
\description{
This function create a plot based on protein level dimension reduction result.
Within the plot, Each cell was colored based on the protein level
}
\examples{
plotProteinFeature(my.obj, "CD11b", protein.reduceDim)
}
