% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{plotVariantFeature}
\alias{plotVariantFeature}
\title{Plot specific variant VAF levels}
\usage{
plotVariantFeature(optima.obj, vaf.name, reduceDim.obj)
}
\arguments{
\item{optima.obj}{optima object.}

\item{vaf.name}{the specific variant name user interested}

\item{reduceDim.obj}{dimension reduction result returned by reduceDim() function.}
}
\value{
A scatter plot based on dimension reduction result. Each cell is colored
based on the protein level. The more expression.
}
\description{
This function create a plot based on VAF dimension reduction result.
Within the plot, Each cell was colored based on the VAF
}
\examples{
plotVariantFeature(my.obj, "chr4:106190862:T/C", vaf.reduceDim)
}
