% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readHdf5}
\alias{readHdf5}
\title{H5 file to optima object function}
\usage{
readHdf5(directory)
}
\arguments{
\item{directory}{Directory for the input h5 file .}
}
\value{
optima object
}
\description{
This function read in a h5 file and return one optima object. The
h5 file can be found in the Tapestri pipeline software output. The .h5
file contains all necessary data needed for single cell DNA and protein
analysis.
}
\examples{
readHdf5("path/to/my/file.h5")
}
\keyword{directory}
