% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\name{reduceDim}
\alias{reduceDim}
\title{Dimension reduction function.}
\usage{
reduceDim(input.mtx)
}
\arguments{
\item{input.mtx}{Input optima object}
}
\value{
List containing PCA result and umap result derived from first 5 PCs
}
\description{
This function reduces dimensions for a data matrix, such data matrix
can be protein or DNA matrix in an optima object.
}
\examples{
reduceDim(example.matrix)
}
