% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{calculatePloidy}
\alias{calculatePloidy}
\title{CNV ploidy calculation function}
\usage{
calculatePloidy(optima.obj, diploid.cell)
}
\arguments{
\item{optima.obj}{optima object.}

\item{diploid.cell}{Cell type that should be considered as diploid cell}
}
\value{
optima object with normalized CNV
}
\description{
The function uses the normalized CNV matrix to calculate the ploidy for each CNV locus.
}
\examples{
calculatePloidy(optima.obj)
}
\keyword{being}
\keyword{optima.obj}
\keyword{ploidy.mtx}
\keyword{updated}
\keyword{with}
