% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\name{drawHeatmap}
\alias{drawHeatmap}
\title{Heatmap function}
\usage{
drawHeatmap(optima.obj, omic.type)
}
\arguments{
\item{optima.obj}{optima object.}

\item{omic.type}{Type of data for heat map. Potential values "dna" and "protein".}
}
\value{
Heat map visualization.
}
\description{
This function creates a heatmap using matrix data, such matrix
data can be DNA or protein.
}
\examples{
drawHeatMap()
}
