% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{findSignature}
\alias{findSignature}
\title{Identify signature protein function}
\usage{
findSignature(optima.obj, cell.type)
}
\arguments{
\item{optima.obj}{optima object.}

\item{cell.type}{Input cell type to compare protein level to all other cell types.}
}
\value{
Data frame of all proteins p-values comparing protein levels of input
cell type to all other cell types.
}
\description{
This function compares protein levels for a input cell type against all other
cells using t test. This function returns a data frame ranked by
FDR adjusted p-value.
}
\examples{
findSignature(optima.obj, cell.type)
}
\keyword{cell.type}
\keyword{optima.obj,}
