% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{normalizeProtein}
\alias{normalizeProtein}
\title{Protein matrix normalization}
\usage{
normalizeProtein(optima.obj)
}
\arguments{
\item{optima.obj}{optima object}
}
\value{
An optima object with protein matrix being normalized and
protein.normalize.method label updated to "normalized"
}
\description{
The function normalizes the protein matrix within an optima object using
CLR method.
}
\examples{
normalizeProtein(optima.obj)
}
\keyword{optima.obj}
