% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\docType{class}
\name{optima-class}
\alias{optima-class}
\alias{optima}
\title{optima object}
\value{
Object containing DNA, protein and CNV single cell sequencing data.
}
\description{
An optima object contains DNA, protein and CNV for Tapestri platform
single cell sequencing data.
}
\section{Slots}{

\describe{
\item{\code{meta.data}}{user defined metadata can be kept with the object}

\item{\code{cell.ids}}{a vector of cell IDs/barcodes from Tapestri. This
vector should contain unique IDs}

\item{\code{cell.labels}}{a vector that is used to store the cell type information
for each cell}

\item{\code{variants}}{a vector of variant IDs}

\item{\code{variant.filter}}{a string that keeps track of if optima object is being QC filtered
on its variant matrix}

\item{\code{vaf.mtx}}{variant matrix}

\item{\code{gt.mtx}}{genotype matrix}

\item{\code{dp.mtx}}{sequencing depth matrix}

\item{\code{gq.mtx}}{genotype quality}

\item{\code{amps}}{a vector of CNV locus}

\item{\code{amp.normalized.method}}{a string that keeps track of if optima object is being normalized
on its CNV matrix}

\item{\code{amp.mtx}}{CNV matrix}

\item{\code{ploidy.mtx}}{ploidy matrix}

\item{\code{proteins}}{a vector of surface protein id}

\item{\code{protein.normalize.method}}{a string that keeps track of if optima object is being normalized
on its protein matrix}

\item{\code{protein.mtx}}{protein matrix}
}}

\examples{
setClass()
}
